moved windows to their own subfolder
