+++ src/ui/windows/researchwindow.cpp
